Fix for the UI debug text and console positions when using LMS scaling.
I believe this should be fixed somewhere else, as fixing it here I'm just providing false dimensions for the viewport.
Additionally, this is also an issue with ScreenPercent, which can't be fixed in the same way, and needs to be fixed probably when the DebugCanvas is created.
I tried it but it didn't seem to do anything.