- Fix for case where people try to take an automation screenshot during BeginPlay.
Now perform the pre-screenshot flush of loading on the first tick of the latent action and not in its constructor. This addresses cases where BP's attempt to take a screenshot in their BeginPlay event while the world is being created.

- Fix for crash that can happen when a window needs resized after a screenshot is captured.

Don't delete (which leads to restoring the viewport size) on response to the screenshot delegates. Instead queue it by one frame


#jira UE-91269
#rb na
#lockdown Cristina.Riveron

[CL 12788643 by Andrew Grant in 4.25 branch]