#jira none
Fix for issue where the Oscilloscope NDI not working in PIE when Create New Audio Device For PIE is enabled in the Editor Settings. Issue was that the fix for the Oscilloscope only using half the buffer wasn't properly set up for FNiagaraDataInterfaceProxyOscilloscope::OnNewDeviceCreated the same way it is in FNiagaraDataInterfaceProxyOscilloscope::RegisterToAllAudioDevices(). To prevent future errors like this, I'm going to consolidate these to one function in Main.

#fyi dan.reynolds, phil.popp
#rb none
#lockdown nick.whiting

[CL 12908966 by Ethan Geller in 4.25 branch]