#jira UE-93033
Merging //Fortnite/Main/Engine/Source/... to //UE4/Release-4.25/Engine/Source/...

Bug fixed: When invalid layout, the editor was crashing. Fixed.

Technical details: When invalid primary area in a layout (e.g., when it contained an invalid tab), FTabManager::RestoreFrom was returning nullptr, that was later casted to SharedRef (causing the crash). Instead, the editor will reset the layout and restart itself (warning the user with a popup). We still maintain an ensure(false), so we can keep individually finding buggy layouts.

Also added some sanity checks to detect earlier this kind of errors.

#rb Chris.Gagnon, Matt.Kuhlenschmidt
#lockdown cristina.riveron

#ROBOMERGE-SOURCE: CL 13431468 in //UE4/Release-4.25/...
#ROBOMERGE-BOT: RELEASE (Release-4.25 -> Release-4.25Plus) (v696-13331060)

[CL 13431526 by gines hidalgo in 4.25-Plus branch]