Fix crash when redoing a create actor action and an edit mesh action at the same time.
Replaying the create action for a static mesh actor places a proxy in the list of primitives to add to the scene. If a subsequent edit mesh action is replayed before the end of the frame, it destroys the render data, then makes a recreate render data context which calls UpdatePrimitiveSceneInfos. This will find the newly added proxy for the mesh we're editing and try to process it, which crashes because the render data is missing. We need to process the list of added primitives immediately after replaying the create action, so that the proxy is added before the render data is destroyed.

#jira UE-89244
#rnx
#rb Francis.Hurteau
#lockdown marcus.wassmer

#ROBOMERGE-SOURCE: CL 13961868 in //UE4/Release-4.25/...
#ROBOMERGE-BOT: RELEASE (Release-4.25 -> Release-4.25Plus) (v720-14121032)

[CL 14149282 by mihnea balta in 4.25-Plus branch]