Sony symbol file refactor  - Merged symbols into a single .uesym file which is kept in the Binaries folder, alongside the .self files.  - Refactored and simplified symbol loading logic to use the sceDbg library to both find the symbol file and deal with ASLR.  - Removed bAllowASLR UBT option. We have ASLR enabled on everything now, as the only reason it was disabled was broken symbol resolution.  - Fixed incorrect implementation of FPlatformStackWalk::CaptureStackBacktrace, which was subtracting 4 from each address, and also made it FORCEINLINE in the header.
For symbol resolution to work, the corresponding .uesym file must be present in the same folder as the .self file, with the same name, e.g.:
 - path/to/build/QAGame.self
 - path/to/build/QAGame.uesym

These files are generated as a post-build step. They are not generated for shipping builds, as the sceDbg library required to workaround ASLR cannot be linked to shipping builds. Callstacks generated in shipping builds will not be annotated with symbol names, and therefore the PLATFORM_SUPPORTS_STACK_SYMBOLS define is 0 in shipping builds.

All other build configurations will work automatically, .e.g via Visual Studio / installed package / deployed staged build etc. The uesym is included in debug, dev and test packages, where it is remapped to /app0/eboot.uesym to match the /app0/eboot.bin executable.

#rb Jeff.Newquist
#jira UE-95768
#jira UE-96748
#jira UE-35901


#ROBOMERGE-OWNER: Luke.Thatcher
#ROBOMERGE-AUTHOR: luke.thatcher
#ROBOMERGE-SOURCE: CL 14336237 via CL 14336248 via CL 14336253
#ROBOMERGE-BOT: (v738-14305781)
#ROBOMERGE[NEXTGEN]: -Release-4.26

[CL 14336337 by Luke Thatcher in 4.25-Plus branch]