USD: Refer to USD root layer with relative paths.
#jira UE-103489
#rb none

Hopefully this is OK: If the usd file is inside the project, the path will be relative like "Content/file.usda". If it's outside the project (or on another drive), it will become an absolute path.

This is as consistent as I can get it without making a mess and overwriting RootLayer manually inside the property event or something like that. I don't really like this change because these meta tags are kind of buggy:
- The "RelativePath" meta tag makes the path relative to the engine binary (no-go because when moving the project to another drive it won't make sense anymore);
- The "RelativeToGameContentDir"  meta tag is flat out broken and has the same effect as RelativePath. Maybe because the comment mentions it's intended for FDirectoryPath instead?
- The "RelativeToGameDir" meta tag sort of works, but:
    - It claims the path will be relative to game dir always, but when using the file picker it really only becomes relative if it's inside the game's *content* folder. Anywhere else inside the game folder and it will become absolute;
    - It does accept programmatically *receiving* paths that are relative to the game (so e.g. "Binaries/file.usda" or "Resources/file.usda"), which is why I do it in USDLayerUtils.cpp;
    - However if the user manually types in a game-relative path like "Binaries/file.usda" (that isn't in the Content folder) it will completely break and turn into something invalid like "E:/Work/daniel_coelho_release_4_26/Engine/Binaries/Win64/Binaries/file.usda".

I chose RelativeToGameDir because it's the most consistent one, and wouldn't break the USD stuff if they later fixed RelativeToGameContentDir to actually be relative to game content. Also I think it may not be super valuable to have the project's level have relative paths to things that aren't inside the project anyway, because if you move the project with respect to them things may break?

#ROBOMERGE-SOURCE: CL 14888471 in //UE4/Release-4.26/...
#ROBOMERGE-BOT: RELEASE (Release-4.26 -> Release-4.26-Chaos) (v763-14813663)

[CL 14888490 by daniel coelho in 4.26-Chaos branch]