Fix for accumulation of results when mulltiple tests using the editor are run on the same builder.
Gauntlet uses a temp path for the /Saved directory to ensure that only artifacts from the current test are copied archived to the network. There is code in the different desktop implementations (TargetDeviceWindows, TargetDeviceMac etc) to clean that folder before Unreal is launched it is only called for cooked builds

Thus when the editor is involved and is receiving screenshots from a device, all of those screenshots accumulate and eventually the four UI tests result in gigabytes of old images being copied to the network shares.

This change ensures every device is given a path that is unique and clean when it is created. Devices will continue to create their own temp-saved path under this folder

For a future UE release [at]Jerome.Delattre is going to take a look for a future release about how we can best remove the duplication of this pattern across all the devices that use temp paths for storing/retrieving artiracts.

[at]jerome.delattre [at]clayton.langford [at]ben.salem [at]cristina.riveron [at]nick.bullard
#jira UE-107468
#rb swarm
#lockdown cristina.riveron

#ROBOMERGE-SOURCE: CL 15282035 in //UE4/Release-4.26/...
#ROBOMERGE-BOT: RELEASE (Release-4.26 -> Release-4.26-Chaos) (v771-15082668)

[CL 15282054 by andrew grant in 4.26-Chaos branch]