Fix crash when reloading the global shader map.
Deleting and reloading the global shader map from the DDC was deleting shaders from underneath the renderer.  Ensure that the cache shaders are refreshed.  See CL# 14368615 for a similar fix.  Breaking out code to refresh primitives with global shaders into a function that can be called from the different pathways global shaders are compiled. Adding this call to refresh in CompileGlobalShaderMap() in situations where we are creating the global shader map.

#jira UE-105177
#rb Arciel.Rekman, Ben.Ingram

[CL 15387473 by Jason Nadro in ue5-early-access branch]