Fix crash when loading a map with many BP render targets on Oculus Quest.
The backbuffer is destroyed and recreated when reloading the map. This means that the FVulkanTexture2D objects which wrap the Oculus-owned backbuffer images are added to the deferred deletion queue. The Vulkan images are destroyed immediately by the Oculus layer, so the driver is free to reuse the handles on subsequent texture allocations. If the new map has many Blueprint render targets, there's a high chance one of them will get a Vulkan image with the same handle as a destroyed backbuffer. When the FVulkanTexture2D objects were eventually deleted, they called NotifyDeletedRenderTarget on the handles without checking if they owned the images, so the handles which were currently used by the BP render targets got incorrectly removed from the layout manager.

Also, NotifyDeletedRenderTarget was not called when the memory manager was asked to delete resources immediately, which could in theory lead to problems when deleting viewports.

This change is deadended because NotifyDeletedRenderTarget was reworked in Main, so the fix must be applied differently there.

#rnx
#jira none
#rb Rolando.Caloca, Steve.Smith
#lockdown Marcus.Wassmer

#ROBOMERGE-SOURCE: CL 15420463 in //UE4/Release-4.26/...
#ROBOMERGE-BOT: RELEASE (Release-4.26 -> Release-4.26-Chaos) (v771-15082668)
#ROBOMERGE[RELEASE]: -Release-Engine-Staging

[CL 15420569 by mihnea balta in 4.26-Chaos branch]