Add translucency flag to Lumen minimal payload
Added bIsTranslucent parameter to FRayTracingMeshCommand. This value is bit-packed into the hit-group user data so that it can be queried upon traversing with the Lumen minimal payload. Encountering an object with translucent blend-mode will cause the Lumen surface-cache lighting mode to retrace the subsequent ray until it reaches a non-translucent object or the environment. Packing the bit into the user data section allows skipping objects during with translucent blend-mode without requiring evaluation of the full material.

An alternative approach would be to split an instance into two separate instances based on translucent blend-mode. Then we could depend on the instance mask to skip translucent objects during traversal.

#rb daniel.wright, krzysztof.narkowicz

[CL 15511114 by Patrick Kelly in ue5-early-access branch]