[Nanite DDC Bump!]
Nanite SceneProxy was not initializing materials correctly based off the modified indices from the section info. This meant that meshes with reordered material slots were rendering incorrectly.

Fix for incorrectly rearranged/truncated/missing material slots for static mesh LOD0 when Nanite enabled and the coarse percentage is < 100% (decimation occuring). Through decimation it's possible for entire static mesh sections to be removed with no assigned triangle ranges (since LOD0 is replaced by a lower LOD vs. the Nanite input data), but the Nanite rendering and editor UI still expects valid slots in the right order for all materials. The fix is to leave the original section array in the same order, but update the entries to match decimation, and leave 0 triangle sections in place for fully removed materials. Certain critical parts of the engine code now also have early-outs to avoid 0 sized draws or collision data.

Fixed incorrect Nanite material rendering when there is an invalid material entry in the list

#rb richard.talbotwatkin

#ushell-cherrypick of 15491868 by Richard.TalbotWatkin
#ushell-cherrypick of 15498869 by graham.wihlidal
#ushell-cherrypick of 15498957 by graham.wihlidal

[CL 15567536 by graham wihlidal in ue5-early-access branch]