Sequencer: Fix copy/paste bug where copy/pasting a possessable binding no longer allows you to paste it again.
This bug was introduced when duplicating a possessable binding was changed to also duplicate the possessed actor. Duplicating the possesssed actor edits the clipboard, which is not desired since the user didn't explicitly invoke that copy operation. So, instead, restore the copied contents of the clipboard after duplicating the possessed actor.

#jira UE-110102
#lockdown cristina.riveron
#rb matt.hoffman

#ROBOMERGE-SOURCE: CL 15630762 in //UE4/Release-4.26/...
#ROBOMERGE-BOT: RELEASE (Release-4.26 -> Release-4.26-Chaos) (v777-15581079)

[CL 15630793 by max chen in 4.26-Chaos branch]