Optimized sub-object look-up within transactions
FPersistentObjectRef would perform a StaticFindObject for each sub-object in its object chain, for each request to resolve the object pointer. This could be quite costly when finding the record for a large transaction (eg, when moving thousands of objects).

This change stores a cache of the resolved object chain, and will use this cache if every object in the chain still passes its validation check. In order to do this, the guts of FPersistentObjectRef have been made private, and the -> operator was also removed as it could promote writing code that would repeatedly resolve the object pointer.

#rb Francis.Hurteau

[CL 15813398 by Jamie Dale in ue5-early-access branch]