Fix broken shadows on Vulkan.
Vulkan was recording buffer upload commands to separate command buffers which were submitted immediately. This meant that a sequence of copy / dispatch, copy / dispatch commands from the high level code would actually execute as copy, copy, dispatch, dispatch. If the two copies reused the same buffer, this led to the first dispatch using data meant for the second dispatch. The fix is to simply use the active command buffer for the copy commands, but we also need to ensure that the high level code doesn't try to do this during a render pass, since that's not allowed in VK. Added validation for this.

#jira UE-108926
#rnx
#rb Luke.Thatcher
#lockdown Michal.Valient

[CL 15930027 by mihnea balta in ue5-early-access branch]