Sequencer: Fix incorrect behaviour when keyframing additive sections.
The problem was that most track editors were using an FSystemInterrogator with no properly allocated channels. This meant that the interrogation had no knowledge of the bound objects or their pre-animated values. As such, any additive keys were not evaluated or decomposed properly, which caused some keyframing situations to produce the wrong result. Now we allocate the channels whenever possible, passing the bound object pointer to it.

#jira UE-118777
#rb max.chen
#lockdown Cristina.Riveron

#ROBOMERGE-SOURCE: CL 16776111 via CL 16776135 via CL 16776155 via CL 16776175
#ROBOMERGE-BOT: RELEASE (Release-4.27-Chaos -> Main) (v836-16769935)
#ROBOMERGE[bot4]: -BOT4-Release-Engine-Staging

[CL 16776181 by ludovic chabant in Main branch]