PatternTool: Fix issues with input fields in the pattern tool.
Extent, radius, and step size can no longer be negative.

Start/end angle are UI limited to the range [-360, 360] but can accept any values. Very large values are useful to create spirals and it would feel a little bit arbitrary to just choose a hard range. Angle shift is clamped to [-180, 180] because values outside of that range are redundant.

Count, CountX, and CountY are clamped to a minimum of 1 with a UI max of 25. The behavior here is unchanged but several redundant UIMins were removed because ClampMin was also already set.

#jira UE-163642
#rb Semion.Piskarev
#preflight 6323592d63312bbbbb07738c

[CL 22032785 by Nickolas Drake in 5.1 branch]