Sequencer: Snapping for key bars
Moved snap to interval to the snap field which is enabled with SetSnapToInterval().  Different operations will enable it depending on their use case. For example, dragging key bars, should have snap to interval enabled if the snap key time to interval setting is enabled. Dragging sections should have snap to interval enabled if the snap section time to interval setting is enabled.

#jira UE-164043
#preflight 632549aeea1c94f4c501a3f5
#rb andrew.rodham

[CL 22075313 by Max Chen in 5.1 branch]