Sequencer: In AddKeysToSection, call ApplyDefault before keys are added.
The default value is a value for the channel when there are no keyframes. For example, if you add keys and then delete them all, the default value is the value of the channel. In the implementation of ApplyDefault, all the setters check that the default value is only set when there are NO keyframes. So, ApplyDefault needs to be called here in AddKeysToSection BEFORE any keys are added.

#jira UE-164765
#rb andrew.rodham, matt.hoffman
#preflight 632cf3e0d51603c54861d0fb

[CL 22158890 by Max Chen in 5.1 branch]