MGPU:  Support for specifying GPU mask for RDG textures, to facilitate a hack fix for GPU Lightmass MGPU transition validation assert.
The assert is due to an architectural bug in RDG where resource transitions for a given resource occur at an arbitrary time in the graph execution, due to batching.  The GPU mask can vary, based on which Pass is executing when a transition for a given resource occurs.  This leads to the internal state used for validation being inconsistent on one of the GPUs.  The hack fix is to make it so the offending resource (PathTracer.LightGrid) only exists on one GPU.  This works because the inconsistent transition state happens on a GPU that the resource isn't actually used on, so removing it from that GPU prevents the assert.  The RDG transition bug doesn't come up often, because most graphs run all-GPU work first (GPU scene update), then run to termination on a single mask, but GPU Lightmass switches back and forth.

The longer term plan (for 5.2) is to scope ALL resources so they only exist on the GPUs they are needed on, then make it so transitions ignore the active GPU mask, and instead apply transitions for all GPUs where a given resource is present.  So there will never be a mismatch in transition state across GPUs.  Scoping resources to specific GPUs saves GPU memory and the CPU performance cost of maintaining the resources, so it would be worth doing on its own.  And it provides other advantages in terms of improved validation and potential automation of cross GPU transfers through RDG.  But this was too big a refactor to do for 5.1, especially with Parallel Rendering work in flight, so I went with the hack fix to avoid the crash.

#jira UE-157789
#rb zach.bethel mihnea.balta
#rnx
#preflight 6339cb1df76de2d4d531462b

[CL 22300686 by jason hoerner in 5.1 branch]