Improved User Parameters Tab. Renaming via Context Menu Deleting via Context Menu
Refactoring of the user param details code so that it is easier to maintain & improve

Removed "User" section in Palette tab

Added "OnParameterRenamed" editor only delegate to the param store so that dependent code can rely on it. The dichtomy between the "ParameterPANELViewModel" handling both parameter internal logic & handling UI interferes with the logic the User Param Tab is supposed to handle. i.e. "ParameterPanelViewModel->RenameParameter" will also attempt to select the parameter we just renamed in the parameter palette tab, so we can't really use that to handle our renaming.

Clicking "user parameters" in the overview graph now summons the user parameter tab
#jira UE-164783
#preflight 63497348ce524ed356e27450

[CL 22527578 by Mateo Egey in 5.1 branch]