Fix for slow performance when applying a new material in a large scene.  The function UWorld::PropagateLightingScenarioChange was calling PropagateLightingScenarioChange on each component, and several types of components were in turn individually creating an FComponentRecreateRenderStateContext, which generates an UpdateAllPrimitiveSceneInfos command for the entire scene.  So it's more or less an O(N^2) operation on the number of primitives in the scene.
Added a feature to batch all the render state recreate operations together using the FGlobalComponentRecreateRenderStateContext class.  Because this is for a hot fix build, the public virtual API for PropagateLightingScenarioChange can't change, so a utility function was added to the FGlobalComponentRecreateRenderStateContext class, to accumulate batched operations in a global array, to be flushed at the end of the operations.  For some components, certain work needs to occur in the scope of the FGlobalComponentRecreateRenderStateContext (between when the render state is destroyed and re-created), so a callback can be provided to do the in-between re-create work.

#jira UE-169342
#rnx
#rb mihnea.balta jason.nadro
#preflight 637b6b078b12eb83a762eb0c
#lockdown mihnea.balta

[CL 23303267 by jason hoerner in 5.1 branch]