OpenXR: pipeline background layer visibility state
Before, we'd just check the background layer visibility from the layer manager across all threads.
However, it's only safe to access on the game thread. When we made the background layer invisible, we
could accidentally disable submission of a frame in the RHI thread.

Now we pipeline background layer visibility, _and_ whether we should submit the layer. If we are using layer emulation, we still want to submit the background layer, since the emulation will composite directly into that layer.

#jira UE-171770
#preflight 63901c6d9549ddaa28a346c7
#rb Jeff.Fisher, Christopher.Fiala
#p4v-cherrypick 23397960

[CL 23433229 by robert srinivasiah in 5.1 branch]