Revert to the old input event for WidgetInteraction components to make sure that the Pointer Event is constructed with the correct Virtual User index.
This fixes an issue where there could be large input deltas on mobile when using a virtual slate user because the index of 8 would not be cleared, since it is not registered with the device mapper.

This was fixed PROPERLY in 5.2 with 22986301, but we can't modify header files in hotfixes. This will fix the incorrect behavior for any users experiencing issues in the live 5.1 release. For this reason, we dont want this to robomerge back up to UE5/Main, hence the -Main tag.


#preflight 639780f035203bc7aa0561e6
#jira FORT-508559
#jira UE-161217
#rb david.garza
#lockdown julien.marchand

[CL 23526388 by ben hoffman in 5.1 branch]